+++ src/ConsoleApplication.fs
@@ -4,7 +4,8 @@ namespace MF.ConsoleApplication
 module MFConsoleApplication =
     open MF.ConsoleStyle
     open OptionsOperators
-    open ResultOperators
+    open MF.ErrorHandling
+    open MF.ErrorHandling.Result.Operators
 
     let consoleApplication =
         let showError application = Error.show application None
@@ -51,11 +52,11 @@ module MFConsoleApplication =
 
     /// Map error by appending a current command.
     let private (<!!*>) result (currentCommand: CurrentCommand) =
-        result <!!> fun __ -> (__, currentCommand)
+        result <@> fun __ -> (__, currentCommand)
 
     /// Map error with appended current command.
     let private (<!!!>) result f =
-        result <!!> fun (error, (__: CurrentCommand)) -> (error |> f, __)
+        result <@> fun (error, (__: CurrentCommand)) -> (error |> f, __)
 
     type private Args = InputValue []
 
@@ -83,7 +84,7 @@ module MFConsoleApplication =
         let (|HasOption|_|) option (args: Args) =
             args |> Array.tryFind (Option.isMatching option)
 
-        let parse applicationOptions (commands: Commands): Args -> Result<Input * UnfilledArgumentDefinitions, ArgsError * CurrentCommand> =
+        let parse output applicationOptions (commands: Commands): Args -> Result<Input * UnfilledArgumentDefinitions, ArgsError * CurrentCommand> =
             fun args ->
                 match args |> List.ofArray with
                 | [] -> Ok (Input.empty, [])
@@ -93,13 +94,13 @@ module MFConsoleApplication =
 
                         let! rawCommandName =
                             rawArg
-                            |> CommandName.createInRuntime <!!> ArgsError.CommandNameError <!!*> currentCommand
+                            |> CommandName.createInRuntime <@> ArgsError.CommandNameError <!!*> currentCommand
 
                         let! (commandName, command) =
                             match commands |> Commands.find rawCommandName with
                             | ExactlyOne (commandName, command) -> Ok (commandName, command)
-                            | MoreThanOne (givenName, names) -> Error (ArgsError.AmbigousCommandFound (givenName, names))
-                            | NoCommand unknownName -> Error (ArgsError.CommandNotFound unknownName)
+                            | MoreThanOne (givenName, names) -> Result.Error (ArgsError.AmbigousCommandFound (givenName, names))
+                            | NoCommand unknownName -> Result.Error (ArgsError.CommandNotFound unknownName)
                             <!!*> currentCommand
 
                         let currentCommand: CurrentCommand = Some (commandName, command)
@@ -108,7 +109,7 @@ module MFConsoleApplication =
 
                         let! parsedInput =
                             rawArgs
-                            |> Input.parse command.Arguments definitions ParsedInput.empty <!!> ArgsError.InputError <!!*> currentCommand
+                            |> Input.parse output command.Arguments definitions ParsedInput.empty <@> ArgsError.InputError <!!*> currentCommand
 
                         let input = {
                             Arguments = parsedInput.Arguments.Add(ArgumentNames.Command, ArgumentValue.Required (commandName |> CommandName.value))
@@ -129,14 +130,14 @@ module MFConsoleApplication =
                 let output = parts.Output
 
                 match args with
-                | Args.ContainsOption OptionsDefinitions.quiet -> output.SetVerbosity Verbosity.Quiet
+                | Args.ContainsOption OptionsDefinitions.quiet -> output.Verbosity <- Verbosity.Quiet
                 | Args.HasOption OptionsDefinitions.verbose verbosity ->
-                    match verbosity with
-                    // todo <later> - add -vvvv special debug mode, which will allow debuging of console-application directly - allow `debug` function
-                    | "-vvv" -> Verbosity.Debug
-                    | "-vv" -> Verbosity.VeryVerbose
-                    | _ -> Verbosity.Verbose
-                    |> output.SetVerbosity
+                    output.Verbosity <-
+                        match verbosity with
+                        // todo <later> - add -vvvv special debug mode, which will allow debuging of console-application directly - allow `debug` function
+                        | "-vvv" -> Verbosity.Debug
+                        | "-vv" -> Verbosity.VeryVerbose
+                        | _ -> Verbosity.Verbose
                 | _ -> ()
 
                 let args =
@@ -162,9 +163,9 @@ module MFConsoleApplication =
                             |> Help.showForCommand output parts.OptionDecorationLevel parts.ApplicationOptions commandName
 
                             Ok ExitCode.Success
-                        | MoreThanOne (givenName, names) -> Error (ArgsError.AmbigousCommandFound (givenName, names))
-                        | NoCommand unknownName -> Error (ArgsError.CommandNotFound unknownName)
-                        <!!> ConsoleApplicationError.ArgsError
+                        | MoreThanOne (givenName, names) -> Result.Error (ArgsError.AmbigousCommandFound (givenName, names))
+                        | NoCommand unknownName -> Result.Error (ArgsError.CommandNotFound unknownName)
+                        <@> ConsoleApplicationError.ArgsError
                         <!!*> currentCommand
                 | Args.ContainsOption OptionsDefinitions.version ->
                     { parts with ApplicationInfo = ApplicationInfo.OnlyNameAndVersion }
@@ -176,7 +177,7 @@ module MFConsoleApplication =
 
                     let! (input, unfilledArguments) =
                         args
-                        |> Args.parse parts.ApplicationOptions parts.Commands <!!!> ConsoleApplicationError.ArgsError
+                        |> Args.parse output parts.ApplicationOptions parts.Commands <!!!> ConsoleApplicationError.ArgsError
 
                     let commandName = input |> Input.getCommandName
                     let command = parts.Commands.[commandName]
@@ -197,20 +198,20 @@ module MFConsoleApplication =
 
                         let! input =
                             input
-                            |> Input.prepareUnfilledArguments unfilledArguments <!!> (ArgsError.InputError >> ConsoleApplicationError.ArgsError) <!!*> currentCommand
+                            |> Input.prepareUnfilledArguments unfilledArguments <@> (ArgsError.InputError >> ConsoleApplicationError.ArgsError) <!!*> currentCommand
 
                         return
                             (input, output)
                             |> command.Execute
                     with
                     | e ->
-                        return! Error (ConsoleApplicationError.ConsoleApplicationError e.Message) <!!*> currentCommand
+                        return! Result.Error (ConsoleApplicationError.ConsoleApplicationError e.Message) <!!*> currentCommand
             }
-        | Error error -> Error (error, currentCommand)
+        | Result.Error error -> Result.Error (error, currentCommand)
 
     let runResult args application =
         application
-        |> runApplication args <!!> fst
+        |> runApplication args <@> fst
 
     let run args application =
         application
@@ -222,6 +223,11 @@ module MFConsoleApplication =
         let args = args |> List.ofArray
         let mutable shouldRun = true
         let mutable exitCode = 0
+
+        let output =
+            application
+            |> ConsoleApplication.output
+            |> Option.defaultValue Output.defaults
 
         let interactiveApplication =
             application
@@ -237,7 +243,7 @@ module MFConsoleApplication =
                 |> Commands.showAvailable parts.Output
             )
 
-            match Console.ask "Command:" with
+            match output.Ask "Command:" with
             | CommandNames.Exit -> shouldRun <- false
             | command ->
                 match run ((command :: args) |> List.toArray) interactiveApplication with