+++ src/Utils.fs
@@ -1,6 +1,8 @@
 namespace MF.ConsoleApplication
 
 module internal OptionsOperators =
+    open MF.ErrorHandling
+
     /// Default value - if value is None, default value will be used
     let (<?=>) defaultValue opt = Option.defaultValue opt defaultValue
 
@@ -27,9 +29,9 @@ module internal OptionsOperators =
 
 [<AutoOpen>]
 module internal Utils =
-    let debug message =
-        if MF.ConsoleStyle.Console.isDebug() then
-            MF.ConsoleStyle.Console.messagef "[DEBUG] %s" message
+    let debug (output: MF.ConsoleStyle.ConsoleStyle) message =
+        if output.IsDebug() then
+            output.Message("<c:gray>[DEBUG]</c> %s", message)
 
     let tee f a =
         f a